SNOMED transitive table and MiADE option (#129)
* SNOMED transitive table and MiADE option

Option to filter if a more specific concept (descendant)
already exists in the patient's record, e.g. if there is already
'Lung cancer' and MiADE detects 'Cancer'. Without this
option, filtering is on exact concept matches only.

The configuration has not been built in and it has not been tested.

* add transitive depulication to annotator config

* fmt: ruff

* remove unused import